+++ server/src/db.rs
@@ -356,23 +356,9 @@ impl Database for Db {
 }
 
 #[cfg(test)]
-mod test_util {
-    use super::*;
-    /// Create a Database with a local store
-    pub fn make_db() -> Db {
-        let name = "test_db";
-        Db::new(
-            DatabaseRules::default(),
-            Some(MutableBufferDb::new(name)),
-            ReadBufferDb::new(),
-            None, // wal buffer
-        )
-    }
-}
-
-#[cfg(test)]
 mod tests {
-    use super::test_util::make_db;
+    use crate::query_tests::utils::make_db;
+
     use super::*;
 
     use arrow_deps::{
@@ -585,341 +571,5 @@ mod tests {
             .collect();
         chunk_ids.sort_unstable();
         chunk_ids
-    }
-}
-
-#[cfg(test)]
-mod test_influxrpc {
-    use query::{
-        exec::{
-            stringset::{IntoStringSet, StringSetRef},
-            Executor,
-        },
-        frontend::influxrpc::InfluxRPCPlanner,
-        predicate::{Predicate, PredicateBuilder},
-    };
-
-    use super::test_scenarios::*;
-
-    /// Creates and loads several database scenarios using the db_setup
-    /// function.
-    ///
-    /// runs table_names(predicate) and compares it to the expected
-    /// output
-    macro_rules! run_table_names_test_case {
-        ($DB_SETUP:expr, $PREDICATE:expr, $EXPECTED_NAMES:expr) => {
-            let predicate = $PREDICATE;
-            for scenario in $DB_SETUP.make().await {
-                let DBScenario { scenario_name, db } = scenario;
-                println!("Running scenario '{}'", scenario_name);
-                println!("Predicate: '{:#?}'", predicate);
-                let planner = InfluxRPCPlanner::new();
-                let executor = Executor::new();
-
-                let plan = planner
-                    .table_names(&db, predicate.clone())
-                    .await
-                    .expect("built plan successfully");
-                let names = executor
-                    .to_string_set(plan)
-                    .await
-                    .expect("converted plan to strings successfully");
-
-                let expected_names = $EXPECTED_NAMES;
-                assert_eq!(
-                    names,
-                    to_stringset(&expected_names),
-                    "Error in  scenario '{}'\n\nexpected:\n{:?}\nactual:\n{:?}",
-                    scenario_name,
-                    expected_names,
-                    names
-                );
-            }
-        };
-    }
-
-    #[tokio::test]
-    async fn list_table_names_no_data_no_pred() {
-        run_table_names_test_case!(NoData {}, empty_predicate(), vec![]);
-    }
-
-    #[tokio::test]
-    async fn list_table_names_no_data_pred() {
-        run_table_names_test_case!(TwoMeasurements {}, empty_predicate(), vec!["cpu", "disk"]);
-    }
-
-    #[tokio::test]
-    async fn list_table_names_data_pred_0_201() {
-        run_table_names_test_case!(TwoMeasurements {}, tsp(0, 201), vec!["cpu", "disk"]);
-    }
-
-    #[tokio::test]
-    async fn list_table_names_data_pred_0_200() {
-        run_table_names_test_case!(TwoMeasurements {}, tsp(0, 200), vec!["cpu"]);
-    }
-
-    #[tokio::test]
-    async fn list_table_names_data_pred_50_101() {
-        run_table_names_test_case!(TwoMeasurements {}, tsp(50, 101), vec!["cpu"]);
-    }
-
-    #[tokio::test]
-    async fn list_table_names_data_pred_250_300() {
-        run_table_names_test_case!(TwoMeasurements {}, tsp(250, 300), vec![]);
-    }
-
-    // No predicate at all
-    fn empty_predicate() -> Predicate {
-        Predicate::default()
-    }
-
-    // make a single timestamp predicate between r1 and r2
-    fn tsp(r1: i64, r2: i64) -> Predicate {
-        PredicateBuilder::default().timestamp_range(r1, r2).build()
-    }
-
-    fn to_stringset(v: &[&str]) -> StringSetRef {
-        v.into_stringset().unwrap()
-    }
-}
-
-#[cfg(test)]
-mod test_table_schema {
-    use arrow_deps::arrow::datatypes::DataType;
-    use data_types::{schema::builder::SchemaBuilder, selection::Selection};
-    use query::{Database, PartitionChunk};
-
-    use super::test_scenarios::*;
-
-    /// Creates and loads several database scenarios using the db_setup
-    /// function.
-    ///
-    /// runs table_schema(predicate) and compares it to the expected
-    /// output
-    macro_rules! run_table_schema_test_case {
-        ($DB_SETUP:expr, $SELECTION:expr, $TABLE_NAME:expr, $EXPECTED_SCHEMA:expr) => {
-            let selection = $SELECTION;
-            let table_name = $TABLE_NAME;
-            let expected_schema = $EXPECTED_SCHEMA;
-
-            for scenario in $DB_SETUP.make().await {
-                let DBScenario { scenario_name, db } = scenario;
-                println!("Running scenario '{}'", scenario_name);
-                println!(
-                    "Getting schema for table '{}', selection {:?}",
-                    table_name, selection
-                );
-
-                // Make sure at least one table has data
-                let mut chunks_with_table = 0;
-
-                for partition_key in db.partition_keys().await.unwrap() {
-                    for chunk in db.chunks(&partition_key).await {
-                        if chunk.has_table(table_name).await {
-                            chunks_with_table += 1;
-                            let actual_schema = chunk
-                                .table_schema(table_name, selection.clone())
-                                .await
-                                .unwrap();
-
-                            assert_eq!(
-                                expected_schema,
-                                actual_schema,
-                                "Mismatch in chunk {}\nExpected:\n{:#?}\nActual:\n{:#?}\n",
-                                chunk.id(),
-                                expected_schema,
-                                actual_schema
-                            );
-                        }
-                    }
-                    assert!(
-                        chunks_with_table > 0,
-                        "Expected at least one chunk to have data, but none did"
-                    );
-                }
-            }
-        };
-    }
-
-    #[tokio::test]
-    async fn list_schema_cpu_all() {
-        // we expect columns to come out in lexographic order by name
-        let expected_schema = SchemaBuilder::new()
-            .tag("region")
-            .timestamp()
-            .field("user", DataType::Float64)
-            .build()
-            .unwrap();
-
-        run_table_schema_test_case!(TwoMeasurements {}, Selection::All, "cpu", expected_schema);
-    }
-
-    #[tokio::test]
-    async fn list_schema_disk_all() {
-        // we expect columns to come out in lexographic order by name
-        let expected_schema = SchemaBuilder::new()
-            .field("bytes", DataType::Int64)
-            .tag("region")
-            .timestamp()
-            .build()
-            .unwrap();
-
-        run_table_schema_test_case!(TwoMeasurements {}, Selection::All, "disk", expected_schema);
-    }
-
-    #[tokio::test]
-    async fn list_schema_cpu_selection() {
-        let expected_schema = SchemaBuilder::new()
-            .field("user", DataType::Float64)
-            .tag("region")
-            .build()
-            .unwrap();
-
-        // Pick an order that is not lexographic
-        let selection = Selection::Some(&["user", "region"]);
-
-        run_table_schema_test_case!(TwoMeasurements {}, selection, "cpu", expected_schema);
-    }
-
-    #[tokio::test]
-    async fn list_schema_disk_selection() {
-        // we expect columns to come out in lexographic order by name
-        let expected_schema = SchemaBuilder::new()
-            .timestamp()
-            .field("bytes", DataType::Int64)
-            .build()
-            .unwrap();
-
-        // Pick an order that is not lexographic
-        let selection = Selection::Some(&["time", "bytes"]);
-
-        run_table_schema_test_case!(TwoMeasurements {}, selection, "disk", expected_schema);
-    }
-}
-
-#[cfg(test)]
-/// This module contains testing scenarios for Db
-mod test_scenarios {
-    use super::*;
-    use query::test::TestLPWriter;
-
-    use super::test_util::make_db;
-
-    /// Holds a database and a description of how its data was configured
-    pub struct DBScenario {
-        pub scenario_name: String,
-        pub db: Db,
-    }
-
-    #[async_trait]
-    pub trait DBSetup {
-        // Create several scenarios, scenario has the same data, but
-        // different physical arrangements (e.g.  the data is in different chunks)
-        async fn make(&self) -> Vec<DBScenario>;
-    }
-
-    /// No data
-    pub struct NoData {}
-    #[async_trait]
-    impl DBSetup for NoData {
-        async fn make(&self) -> Vec<DBScenario> {
-            let partition_key = "1970-01-01T00";
-            let db = make_db();
-            let scenario1 = DBScenario {
-                scenario_name: "New, Empty Database".into(),
-                db,
-            };
-
-            // listing partitions (which may create an entry in a map)
-            // in an empty database
-            let db = make_db();
-            assert_eq!(db.mutable_buffer_chunks(partition_key).await.len(), 1); // only open chunk
-            assert_eq!(db.read_buffer_chunks(partition_key).await.len(), 0);
-            let scenario2 = DBScenario {
-                scenario_name: "New, Empty Database after partitions are listed".into(),
-                db,
-            };
-
-            // a scenario where the database has had data loaded and then deleted
-            let db = make_db();
-            let data = "cpu,region=west user=23.2 100";
-            let mut writer = TestLPWriter::default();
-            writer.write_lp_string(&db, data).await.unwrap();
-            // move data out of open chunk
-            assert_eq!(db.rollover_partition(partition_key).await.unwrap().id(), 0);
-            // drop it
-            db.drop_mutable_buffer_chunk(partition_key, 0)
-                .await
-                .unwrap();
-
-            assert_eq!(db.mutable_buffer_chunks(partition_key).await.len(), 1);
-
-            assert_eq!(db.read_buffer_chunks(partition_key).await.len(), 0); // only open chunk
-
-            let scenario3 = DBScenario {
-                scenario_name: "Empty Database after drop chunk".into(),
-                db,
-            };
-
-            vec![scenario1, scenario2, scenario3]
-        }
-    }
-
-    /// Two measurements data in a single mutable buffer chunk
-    pub struct TwoMeasurements {}
-    #[async_trait]
-    impl DBSetup for TwoMeasurements {
-        async fn make(&self) -> Vec<DBScenario> {
-            let partition_key = "1970-01-01T00";
-            let data = "cpu,region=west user=23.2 100\n\
-                        cpu,region=west user=21.0 150\n\
-                        disk,region=east bytes=99i 200";
-
-            let db = make_db();
-            let mut writer = TestLPWriter::default();
-            writer.write_lp_string(&db, data).await.unwrap();
-            let scenario1 = DBScenario {
-                scenario_name: "Data in open chunk of mutable buffer".into(),
-                db,
-            };
-
-            let db = make_db();
-            let mut writer = TestLPWriter::default();
-            writer.write_lp_string(&db, data).await.unwrap();
-            db.rollover_partition(partition_key).await.unwrap();
-            let scenario2 = DBScenario {
-                scenario_name: "Data in closed chunk of mutable buffer".into(),
-                db,
-            };
-
-            let db = make_db();
-            let mut writer = TestLPWriter::default();
-            writer.write_lp_string(&db, data).await.unwrap();
-            db.rollover_partition(partition_key).await.unwrap();
-            db.load_chunk_to_read_buffer(partition_key, 0)
-                .await
-                .unwrap();
-            let scenario3 = DBScenario {
-                scenario_name: "Data in both read buffer and mutable buffer".into(),
-                db,
-            };
-
-            let db = make_db();
-            let mut writer = TestLPWriter::default();
-            writer.write_lp_string(&db, data).await.unwrap();
-            db.rollover_partition(partition_key).await.unwrap();
-            db.load_chunk_to_read_buffer(partition_key, 0)
-                .await
-                .unwrap();
-            db.drop_mutable_buffer_chunk(partition_key, 0)
-                .await
-                .unwrap();
-            let scenario4 = DBScenario {
-                scenario_name: "Data in only buffer and not mutable buffer".into(),
-                db,
-            };
-
-            vec![scenario1, scenario2, scenario3, scenario4]
-        }
     }
 }
